core-main/src/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
pyyaml==6.0.1
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
email-validator==2.3.0
qrcode[pil]==8.2
apscheduler==3.10.4
certifi==2024.8.30
firebase-admin>=6.0.0
